Move the tensor_fields_container.py to get ready for refactoring.
The chunking and vectorisation logic will be removed from tensor_fields_container, so it makes sense to move it to vespa_index since it's only used by add_documents_handler.py

Also moved all the unit tests for these two files to unit_tests folder.
